requirements.txt
# Core ML libraries
torch>=2.0.0
transformers>=4.35.0
peft>=0.6.0
bitsandbytes>=0.41.0
trl>=0.7.0
datasets>=2.14.0
accelerate>=0.24.0

# Model conversion and inference
llama-cpp-python>=0.2.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Utilities
tqdm>=4.65.0
wandb>=0.15.0
tensorboard>=2.13.0
huggingface-hub>=0.17.0

# Development and testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0 
